+++ src/components/MovementComponent.cpp
@@ -17,7 +17,7 @@ MovementComponent::~MovementComponent()
 void MovementComponent::Update(const float deltaTime)
 {
 	if (m_rotationFollowsVelocity and !Math::IsNearlyEqual(m_velocity.Lenght(), 0.0)) {
-		const float angle = Math::GetAngleFromDirection(m_velocity);
+		const double angle = Math::GetAngleFromDirection(m_velocity);
 		m_parent->m_transform.SetRotation(angle);
 	}
 	AddPositionDelta(m_velocity * m_speed);
